refactor(sidebar): moved retrieving user to Sidebar from ActivitySidebar
