+++ src/elements/content-sidebar/ActivitySidebar.js
@@ -18,7 +18,7 @@ import messages from '../common/messages';
 import SidebarContent from './SidebarContent';
 import { WithAnnotatorContextProps, withAnnotatorContext } from '../common/annotator-context';
 import { EVENT_DATA_READY, EVENT_JS_READY } from '../common/logger/constants';
-import { getBadUserError, getBadItemError } from '../../utils/error';
+import { getBadUserError } from '../../utils/error';
 import { mark } from '../../utils/performance';
 import { withAPIContext } from '../common/api-context';
 import { withErrorBoundary } from '../common/error-boundary';
@@ -61,6 +61,7 @@ type ExternalProps = {
     activeFeedEntryId?: string,
     activeFeedEntryType?: FocusableFeedItemType,
     currentUser?: User,
+    currentUserError?: Errors,
     getUserProfileUrl?: GetProfileUrlCallback,
     hasReplies?: boolean,
     hasTasks?: boolean,
@@ -98,8 +99,6 @@ type State = {
     activityFeedError?: Errors,
     approverSelectorContacts: SelectorItems<UserMini | GroupMini>,
     contactsLoaded?: boolean,
-    currentUser?: User,
-    currentUserError?: Errors,
     feedItems?: FeedItems,
     feedItemsStatusFilter?: FeedItemStatus,
     mentionSelectorContacts?: SelectorItems<UserMini>,
@@ -154,10 +153,7 @@ class ActivitySidebar extends React.PureComponent<Props, State> {
     }
 
     componentDidMount() {
-        const { currentUser } = this.props;
-
         this.fetchFeedItems(true);
-        this.fetchCurrentUser(currentUser);
     }
 
     handleAnnotationDelete = ({ id, permissions }: { id: string, permissions: AnnotationPermission }) => {
@@ -214,32 +210,6 @@ class ActivitySidebar extends React.PureComponent<Props, State> {
     }
 
     /**
-     * Fetches a Users info
-     *
-     * @private
-     * @param {User} [user] - Box User. If missing, gets user that the current token was generated for.
-     * @param {boolean} shouldDestroy
-     * @return {void}
-     */
-    fetchCurrentUser(user?: User, shouldDestroy: boolean = false): void {
-        const { api, file } = this.props;
-
-        if (!file) {
-            throw getBadItemError();
-        }
-
-        if (typeof user === 'undefined') {
-            api.getUsersAPI(shouldDestroy).getUser(
-                file.id,
-                this.fetchCurrentUserSuccessCallback,
-                this.fetchCurrentUserErrorCallback,
-            );
-        } else {
-            this.setState({ currentUser: user, currentUserError: undefined });
-        }
-    }
-
-    /**
      * Success callback for fetching feed items
      */
     feedSuccessCallback = (): void => {
@@ -267,8 +237,7 @@ class ActivitySidebar extends React.PureComponent<Props, State> {
         onSuccess: ?Function,
         onError: ?Function,
     ): void => {
-        const { currentUser } = this.state;
-        const { file, api } = this.props;
+        const { currentUser, file, api } = this.props;
 
         if (!currentUser) {
             throw getBadUserError();
@@ -343,8 +312,7 @@ class ActivitySidebar extends React.PureComponent<Props, State> {
     };
 
     updateTaskAssignment = (taskId: string, taskAssignmentId: string, status: TaskCollabStatus): void => {
-        const { file, api, onTaskAssignmentUpdate } = this.props;
-        const { currentUser = {} } = this.state;
+        const { currentUser, file, api, onTaskAssignmentUpdate } = this.props;
 
         const successCallback = () => {
             this.feedSuccessCallback();
@@ -449,8 +417,7 @@ class ActivitySidebar extends React.PureComponent<Props, State> {
      * @return {void}
      */
     createComment = (text: string, hasMention: boolean): void => {
-        const { file, api, hasReplies, onCommentCreate } = this.props;
-        const { currentUser } = this.state;
+        const { currentUser, file, api, hasReplies, onCommentCreate } = this.props;
 
         if (!currentUser) {
             throw getBadUserError();
@@ -631,17 +598,6 @@ class ActivitySidebar extends React.PureComponent<Props, State> {
     };
 
     /**
-     * User fetch success callback
-     *
-     * @private
-     * @param {Object} currentUser - User info object
-     * @return {void}
-     */
-    fetchCurrentUserSuccessCallback = (currentUser: User): void => {
-        this.setState({ currentUser, currentUserError: undefined });
-    };
-
-    /**
      * File approver contacts fetch success callback
      *
      * @private
@@ -726,29 +682,6 @@ class ActivitySidebar extends React.PureComponent<Props, State> {
             include_uploader_collabs: false,
         });
     }
-
-    /**
-     * Handles a failed file user info fetch
-     *
-     * @private
-     * @param {ElementsXhrError} e - API error
-     * @return {void}
-     */
-    fetchCurrentUserErrorCallback = (e: ElementsXhrError, code: string) => {
-        this.setState({
-            currentUser: undefined,
-            currentUserError: {
-                maskError: {
-                    errorHeader: messages.currentUserErrorHeaderMessage,
-                    errorSubHeader: messages.defaultErrorMaskSubHeaderMessage,
-                },
-            },
-        });
-
-        this.errorCallback(e, code, {
-            error: e,
-        });
-    };
 
     /**
      * Gets the user avatar URL
@@ -874,6 +807,8 @@ class ActivitySidebar extends React.PureComponent<Props, State> {
 
     render() {
         const {
+            currentUser,
+            currentUserError,
             elementId,
             file,
             isDisabled = false,
@@ -883,14 +818,7 @@ class ActivitySidebar extends React.PureComponent<Props, State> {
             activeFeedEntryType,
             onTaskView,
         } = this.props;
-        const {
-            currentUser,
-            approverSelectorContacts,
-            mentionSelectorContacts,
-            contactsLoaded,
-            activityFeedError,
-            currentUserError,
-        } = this.state;
+        const { approverSelectorContacts, mentionSelectorContacts, contactsLoaded, activityFeedError } = this.state;
 
         return (
             <SidebarContent
